+++ packages/platform/src/lib/content/prism/index.ts
@@ -2,10 +2,6 @@ import { PrismHighlighter } from './prism-highlighter.js';
 
 export { PrismHighlighter };
 
-export type WithPrismHighlighterOptions = {
-  additionalLangs?: string[];
-};
-
 export function getPrismHighlighter() {
   return new PrismHighlighter();
 }
