+++ packages/platform/src/lib/content/prism/options.ts
@@ -1,0 +1,3 @@
+export type WithPrismHighlighterOptions = {
+  additionalLangs?: string[];
+};
