+++ packages/platform/src/lib/content/shiki/options.ts
@@ -1,0 +1,14 @@
+import {
+  ShikiHighlighterOptions,
+  ShikiHighlightOptions,
+} from './shiki-highlighter.js';
+
+import { BundledLanguage } from 'shiki/langs';
+
+export interface WithShikiHighlighterOptions {
+  highlighter?: Partial<ShikiHighlighterOptions> & {
+    additionalLangs?: BundledLanguage[];
+  };
+  highlight?: ShikiHighlightOptions;
+  container?: string;
+}
